requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
PyYAML>=6.0
scipy>=1.11.0
tqdm>=4.66.0
Pillow>=10.0.0
torch>=2.2.0
torchvision>=0.17.0
facenet-pytorch>=2.6.0
streamlit>=1.40.0
mediapipe==0.10.14
Flask>=3.0.0
